ci: add CI feature (#43)
* ci: add build and quality check

* ci: add `PR` title check

* ci: add slack notify check

* ci: implement jacoco report generator, sonarqube, pmd and checkstyle lint check

* ci: skip CI for sample app

* ci: add `build check` and `coverage`  badge to readme file

* ci: fix PMD and checkstyle lint plugin version

* test: fix the MSG_MAX_CREATE_SIZE in AnalyticsClient

Decrease the size by 1, otherwise the test was failing

* ci: add suppression rule for checkstyle

* chore: add feat/latestPull to test the working of build and quality check

Once test completes, revert the change.

* chore: disable badge support

Once test completes, revert the change.

* ci: add `Setup` process for `lint` action

* ci: set `sonar.java.skipUnchanged` to false

To try to remove optimise this log message: `[INFO] The Java analyzer is running in a context where unchanged files can be skipped. Full analysis is performed for changed files, optimized analysis for unchanged files.`

* ci: set `source` and `test` directory for sonar analysis

* ci: set the `exclusion` property for sonar

* ci: set `main` and `test` property for sonar

* ci: comment source and test path for

* chore: try different sonar command

* ci: add jacoco version in pom

* chore: revert sonar command

* chore: trigger CI

* chore: remove comment

* ci: retry badge action

It was producing error earlier, retry now to understand the issue.

* ci: change the name of the action to `Jacoco Badge generator`

* ci: try different action to push the Jacoco report badge

* Revert "ci: try different action to push the Jacoco report badge"

This reverts commit 8f509724ef0802700b9c363651bb332d7196ac67.

* ci: try different action to push the Jacoco report badge

* ci: try different action to push the Jacoco report badge

* ci: disable Jacoco badge generator

Having issues (because of EndBug/add-and-commit@v9.1.1) while uploading back the badge to the git branch, hence disabled it. Error log: `Error: Error: fatal: You are not currently on a branch. To push the history leading to the current (detached HEAD) state now, use git push origin HEAD:<name-of-remote-branch>`

* ci: add `Notion PR sync` action

* Revert: remove `feat/latestPull` branch from the action.

This reverts commit: 4d7651f.
As the result of the CI is satisfactory, hence removing the `feat/latestPull`.
This will also disable the CI pipeline for this PR.

* fix: change project name to Java SDK in slack-notify

* ci: remove slack-notify

This is not needed as part of the CI.

* ci(lint): upload lint result to sonarqube
Move the lint to the same job so that lint result gets generated and uploaded to sonarqube